Refactored naming for directories and added debug.log to gitignore
