+++ .gitignore
@@ -21,5 +21,6 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+src/debug.log
 
 .idea/
